requirements.txt
# -----------------------------
# Core Test Framework
# Framework for writing and executing automated tests, supports fixtures, markers, assertions
# -----------------------------
pytest

# -----------------------------
# HTTP Client for API Testing
# For making REST API calls: GET, POST, PUT, DELETE, with headers, auth, payloads
# -----------------------------
requests

# -----------------------------
# Configuration Management
 # Parse YAML config files for environment variables, credentials, URLs, etc.
# -----------------------------
pyyaml

# -----------------------------
# Test Reporting
# Generates HTML reports for test results, supports embedded logs
# -----------------------------
pytest-html

# -----------------------------
# Parallel Test Execution
# Run tests in parallel across multiple workers, supports loadscope/loadfile strategies
# -----------------------------
pytest-xdist


